# src/main/resources/api/api.yml
openapi: 3.0.3
info:
  title: User Management API
  version: 1.0.0
servers:
  - url: http://localhost:8090/autify/api/v1
    description: Optional server description, e.g. Internal staging server for testing
tags:
  - name: user
    description: Operations for user.
  - name: role
    description: Operations for role.
paths:
  /users:
    post:
      tags:
        - user
      operationId: createUser
      summary: Add a new user
      requestBody:
        description: User object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input, user not added
        '500':
          description: Application Internal Error
    put:
      tags:
        - user
      operationId: updateUser
      summary: Update existing user
      requestBody:
        description: User object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags:
        - user
      summary: Get user detail
      description: This can only be done by the logged in user.
      operationId: getUser
      parameters:
        - name: userId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Get User detail successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: User not found.
  /roles:
    post:
      tags:
        - role
      operationId: createRole
      summary: Add a new role
      requestBody:
        description: Role object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        201:
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
    put:
      tags:
        - role
      operationId: updateRole
      summary: Update a existing role
      requestBody:
        description: Role object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        200:
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
  /roles/{roleId}:
    delete:
      tags:
        - role
      summary: Delete role
      description: This can only be done by the logged in user.
      operationId: deleteRole
      parameters:
        - name: roleId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Role not found.
    get:
      tags:
        - role
      summary: Get role detail
      description: This can only be done by the logged in user.
      operationId: getRole
      parameters:
        - name: roleId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Get Role detail successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Role not found.
  /roles/{roleId}/permissions:
    get:
      tags:
        - role
      operationId: getOwnedPermissions
      summary: Get owned permissions of role
      parameters:
        - name: roleId
          in: path
          description: The role id that needs to be get permissions
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Get permissions successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Role not found
        '400':
          description: Invalid input, user not added
        '500':
          description: Application Internal Error
  /roles/{roleId}/permissions/{permissionId}:
    post:
      tags:
        - role
      operationId: assignPermission
      summary: Assign permission for specific role
      parameters:
        - name: roleId
          in: path
          description: The role id that needs to be assigned permission
          required: true
          schema:
            type: integer
            format: int64
        - name: permissionId
          in: path
          description: The permission id that need to be assigned
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Permission assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Permission or Role not found
        '400':
          description: Invalid input, user not added
        '500':
          description: Application Internal Error
    delete:
      tags:
        - role
      operationId: removeAssignedPermission
      summary: Remove Assigned permission from specific role
      parameters:
        - name: roleId
          in: path
          description: The role id that needs to be removed assigned permission
          required: true
          schema:
            type: integer
            format: int64
        - name: permissionId
          in: path
          description: The assigned permission id that need to be removed
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Permission removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Permission or Role not found
        '400':
          description: Invalid input, user not added
        '500':
          description: Application Internal Error
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        permissions:
          type: array
          items:
            - format: string
      required:
        - name